requirements.txt
Flask
tensorflow
numpy
Pillow
kaggle
kagglehub
